+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+v0.2.2:
+  * Add debug and trace logs for skipped noise bytes.
+
 v0.2.1:
   * Change error type of sync write functions to reflect lack of read phase.
   * Add support for the bulk read and write instructions.